# src/Web/ClientApp/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Web | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5257/"
    }
  ],
  "paths": {
...
    "/api/business-names": {
      "get": {
        "tags": [
          "Business Names"
        ],
        "operationId": "GetBusinessNames",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessNameDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Business Names"
        ],
        "operationId": "CreateBusinessName",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessNameCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/business-names/{id}": {
      "put": {
        "tags": [
          "Business Names"
        ],
        "operationId": "UpdateBusinessName",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessNameCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Business Names"
        ],
        "operationId": "DeleteBusinessName",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/business-names/{id}/cancel": {
      "post": {
        "tags": [
          "Business Names"
        ],
        "operationId": "CancelBusinessName",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ato/link/start": {
      "post": {
        "tags": [
          "ATO"
        ],
        "operationId": "StartAtoLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtoLinkStartResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/ato/link/poll": {
      "post": {
        "tags": [
          "ATO"
        ],
        "operationId": "PollAtoLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtoPollResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/ato/link/select": {
      "post": {
        "tags": [
          "ATO"
        ],
        "operationId": "SelectAtoAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtoSelectResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/ato/mark-connected": {
      "post": {
        "tags": [
          "ATO"
        ],
        "operationId": "MarkAtoConnected",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkConnectedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtoAgentDto": {
        "required": [
          "name",
          "abn",
          "ran"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "abn": {
            "type": "string"
          },
          "ran": {
            "type": "string"
          }
        }
      },
      "AtoLinkStartResult": {
        "required": [
          "attemptId",
          "referenceCode"
        ],
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "referenceCode": {
            "type": "string"
          }
        }
      },
      "AtoLinkStatus": {
        "enum": [
          "Pending",
          "Linked",
          "Failed",
          "Expired"
        ]
      },
      "AtoPollResult": {
        "required": [
          "status",
          "agents",
          "reason"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/AtoLinkStatus"
          },
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AtoAgentDto"
            }
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "AtoSelectResult": {
        "required": [
          "connected",
          "syncedCount",
          "nomination"
        ],
        "type": "object",
        "properties": {
          "connected": {
            "type": "boolean"
          },
          "syncedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nomination": {
            "type": "string"
          }
        }
      },
...
      "BusinessNameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "dateRegistered": {
            "type": "string"
          },
          "renewalDate": {
            "type": "string"
          },
          "asicKey": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CancelBody": {
        "required": [
          "scope",
          "cardName",
          "cardNumber",
          "expiry",
          "ccv"
        ],
        "type": "object",
        "properties": {
          "scope": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardName": {
            "type": [
              "null",
              "string"
            ]
          },
          "cardNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "expiry": {
            "type": [
              "null",
              "string"
            ]
          },
          "ccv": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreateBusinessNameCommand": {
        "required": [
          "name",
          "dateRegistered",
          "renewalDate",
          "asicKey"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "dateRegistered": {
            "type": "string"
          },
          "renewalDate": {
            "type": "string"
          },
          "asicKey": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "IdResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "MarkConnectedRequest": {
        "required": [
          "desired"
        ],
        "type": "object",
        "properties": {
          "desired": {
            "type": "boolean"
          }
        }
      },
...
      "PollLinkRequest": {
        "required": [
          "attemptId"
        ],
        "type": "object",
        "properties": {
          "attemptId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "SelectAgentRequest": {
        "required": [
          "abn"
        ],
        "type": "object",
        "properties": {
          "abn": {
            "type": "string"
          }
        }
      },
      "StartLinkRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "UpdateBusinessNameCommand": {
        "required": [
          "id",
          "name",
          "dateRegistered",
          "renewalDate",
          "asicKey"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "dateRegistered": {
            "type": "string"
          },
          "renewalDate": {
            "type": "string"
          },
          "asicKey": {
            "type": "string"
          }
        }
      }
    }
  },
...
}